# swagger/model_user.go
/*
 * openshift basic identity
 *
 * openshift basic identity provider
 *
 * API version: 1.0.0
 * Contact: zhu.xiaowei@99cloud.net
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type User struct {

	Id string `json:"id,omitempty"`

	Username string `json:"username"`

	Password string `json:"password"`

	Email string `json:"email,omitempty"`

	Sub string `json:"sub,omitempty"`

	Name string `json:"name,omitempty"`

	PreferredUsername string `json:"preferred_username,omitempty"`
}